10711 Removed unused pages.
